updated file system to use express router for better separation of concerns
